server/requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.31
httpx>=0.27,<0.28
python-dotenv>=1.0,<2.0
